Align encoding as a core type, and align flag to u8
This will make it consistent with what string expects.
