+++ artichoke-backend/src/encoding_registry.rs
@@ -3,29 +3,17 @@ use bstr::ByteVec;
 use crate::core::EncodingRegistry;
 use crate::def::NotDefinedError;
 use crate::error::Error;
-use crate::extn::core::encoding::Encoding;
+use crate::extn::core::encoding::{Encoding, RUBY_TYPE};
 use crate::ffi::InterpreterExtractError;
 use crate::prelude::*;
 use crate::sys;
 use crate::value::Value;
 use crate::Artichoke;
 use artichoke_core::constant::DefineConstant;
+use artichoke_core::encoding::Encoding as _;
 use artichoke_core::value::Value as _;
 
 pub type Spec = Encoding;
-
-/// Define an encoding which can be stored by `EncodingRegistry`.
-pub trait EncodingSpec {
-    const RUBY_TYPE: &'static str;
-
-    fn id(&self) -> i64;
-    fn aliases(&self) -> Vec<Vec<u8>>;
-    fn is_ascii_compatible(&self) -> bool;
-    fn is_dummy(&self) -> bool;
-    fn inspect(&self) -> &'static str;
-    fn name(&self) -> &'static str;
-    fn names(&self) -> &'static [&'static str];
-}
 
 impl EncodingRegistry for Artichoke {
     type Value = Value;
@@ -40,17 +28,17 @@ impl EncodingRegistry for Artichoke {
             let encoding_class = state
                 .classes
                 .get::<Self::Spec>()
-                .ok_or_else(|| NotDefinedError::class(Spec::RUBY_TYPE))?;
+                .ok_or_else(|| NotDefinedError::class(RUBY_TYPE))?;
 
             let rclass = encoding_class.rclass();
             unsafe { self.with_ffi_boundary(|mrb| rclass.resolve(mrb)) }?
-                .ok_or_else(|| NotDefinedError::class(Spec::RUBY_TYPE))?
+                .ok_or_else(|| NotDefinedError::class(RUBY_TYPE))?
         };
 
         // Allocate the encoding
         // TODO: Sanitize that the encoding doesn't already exist in the registry
-        let id = spec.id();
-        let obj = unsafe { sys::mrb_sys_new_encoding(id) };
+        let id = spec.flag();
+        let obj = unsafe { sys::mrb_sys_new_encoding(i64::from(id)) };
         let encoding = Value::from(obj);
 
         // Generate and assign all the constants for this encoding
@@ -77,19 +65,28 @@ impl EncodingRegistry for Artichoke {
         todo!()
     }
 
+    #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
     fn encoding_for(&mut self, value: &mut Self::Value) -> Result<&Self::Spec, Self::Error> {
         if value.ruby_type() != Ruby::Encoding {
             let mut message = String::from("uninitialized ");
-            message.push_str(Spec::RUBY_TYPE);
+            message.push_str(RUBY_TYPE);
             return Err(TypeError::from(message).into());
         }
 
         let value = unsafe { value.inner().value.i };
+        let value: u8 = if let 0..=255 = value {
+            // SAFETY: 0..=255 will always be able to be cast into a `u8`
+            value as u8
+        } else {
+            let mut message = String::from("uninitialized ");
+            message.push_str(RUBY_TYPE);
+            return Err(TypeError::from(message).into());
+        };
 
         let state = self.state.as_deref_mut().ok_or_else(InterpreterExtractError::new)?;
         state.encodings.get(&value).ok_or_else(|| {
             let mut message = String::from("uninitialized ");
-            message.push_str(Spec::RUBY_TYPE);
+            message.push_str(RUBY_TYPE);
             TypeError::from(message).into()
         })
     }
