+++ artichoke-backend/src/extn/core/encoding/mod.rs
@@ -15,7 +15,7 @@ mod backend;
 pub(in crate::extn) mod mruby;
 pub(super) mod trampoline;
 
-const RUBY_TYPE: &str = "Encoding";
+pub const RUBY_TYPE: &str = "Encoding";
 pub use backend::spinoso::Encoding;
 
 #[cfg(test)]
